Fix missing imports of _
A number of files used _() but didn't import it.  Add the missing
imports.

Note that tox.ini is updated as well to ensure that the hacking rule
about only importing modules is ignored for this case.  This avoids
having to add #noqa to every instance of this import.

Test coverage that hits these code paths is being added in another
review, where this originally came up:
Idd883fda93be4f7d28554c26902da0b9744f0bf1

Closes-bug: #1258254
Change-Id: I19b89be8864ce3a27fef48d5b00e44110fa65adf