Add functionaltests directory
The gate-solum-devstack-dsvm job is now running properly.  It starts up
devstack including solum.  Now we need a suite of tests to run against
solum running in devstack.  This is intended to be the start of that
suite.  The first test is just a request to the root of the API to
verify that version discovery is working (and that we can talk to the
API at all).

Tempest does not have any sort of plugin mechanism, so we can't run in
the context of tempest.  The advice we're given is to just develop our
own functional test suite for now that we can work to integrate later
if/when the project were to become incubated.

Change-Id: Ia6cebf1411a45336ead4442cdc9be4ade4745455