Resource for querying available language packs
We want the ability to specify the language pack that should be used
to build a DU. For this, we need the ability to query the language
packs that are currently available. Defined a resource to query
available language packs. The resource is defined at
(/v1/language_packs).

This code is based on the new API worker architecture

Partly implements: blueprint specify-lang-pack

Change-Id: I32fbfe2388f1e571d8ddcce001b6752d08747e1e