make the development instructions more consistent
Our setup is already more complex than installing the python
package. To run the functional tests we need a devstack setup
so try direct people to install via devstack rather than
the trivial "pip install".

Also installation.rst was not included in the docs as it was not
referenced. This should be under the new section "getting_started".

Change-Id: Ia234506e783db2c40428c226a4d26000ded8a9b1