Standardize Data types accross model objects
Following the discussion on https://review.openstack.org/#/c/69601/
this patch standardize the data types of resources for the model
objects. Link have to be replaced by URI.

In the component model, nested components are represented as a list
of IDs that can then be queried individually. The attribute
abreviated has been added to specify if the component has nested
components at more than one level depth

Change-Id: I58c8f95fb4fb170efb071b81d76bfe8012ad050a