Make sure the alembic schema are run in the tests
We need to make sure that the migrations are run to get
the latest schema when the db is created in the new Database
fixture.

Note: alembic migrations do not seem to work with an in-memory sqlite
as the alembic command is run in a different process
resulting in no schema been applied.

Now we use a new sqlite database (tempfile) for each test and make use of the
NestedTempfile fixture to delete the file.

Change-Id: Iea7f5c90e91404a89bdd526f609f4352c7f5dbd3