Create an API endpoint for Zuul
This patch is the first part of the zuul-endpoint blueprint. It
creates the API endpoint in solum for triggers.

This endpoint has to be public, so this patch also add a wrapper
on keystone AuthProtocol to allow unauthenticated access to all
requests targeting an url starting with '/{api_version}/public/

Partially implements blueprint zuul-endpoint

Change-Id: I0a66303111e163c60a00157f1fb768fb459930bc