Stack attributes were used the wrong way after a stack creation.
The previous commit https://review.openstack.org/#/c/83911/
fixed the circular reference Error by passing the raw file to
the heat client.

Actually, it was the only part that needed to be changed. With the
current code, the stack is created bu after that,we cannot get the
description from the template file.

This patch also removes the "get" call to the heat client. All the
info we need is given by the response of the create call.

Fixes bug #1300121

Change-Id: Ib84e1c0e402e8fb5de085933fa4aca6f33213aad