Add configurable handler selection for Worker
Added to solum-worker the configuration value 'handler'
to select which one of the available modules in its
handlers directory to use as its endpoint. The result is
being able to configure exactly how images will be built.

At present we only have shell.py, which uses subprocess
to run the build-app script, but in the future there may
be other handlers to use perhaps Jenkins or another process.

Since worker already had shell.py, I have added default.py
mainly for contract enforcement, and so there is a default
to fall back upon should shell.py be unsuitable to a user.

Change-Id: Ic9a40cbef111ef3c28c6d629dd4209398d1d52ae