Make deployer sleep time variable.
The deployer waits for a configured amount of time before
querying the Heat stack.
If this wait interval is very small (e.g. 1 second) the number
of requests to the Heat API will be significantly large.
If this wait interval is large (e.g. 5 seconds) then, the
deployer will necessarily have to wait for this period even if
the stack creation completes within this time.
This is a disadvantage if we are using Docker containers since
it takes approximately 1 sec to provision a new docker container.

Added code to make this sleep time variable. Initially we set
it to a small number and then this time keeps on growing by the
'growth_factor' configuration option.

Change-Id: I63fff41a928054addb3bfd44d3bf687a8e6d0e92