Add configurable handler selection for Deployer
Added to solum-deployer the configuration value 'handler'
to select which one of the available modules in its
handlers directory to use as its endpoint. The result is
being able to configure exactly how images will be deployed.

At present we only have default.py, which uses Heat to deploy
a stack for the app, but in the future there may be other
handlers to use another process.

Since default.py was using Heat, I renamed it to heat.py and
added a new default.py mainly for contract enforcement.

Change-Id: I4c2aefc7d60e421c176bb8e854548486e9d800a3