Squash alembic files.
This patch squashes all the alembic migration scripts
into one file.

I used the hash from the most recently added alembic
migration file. This is to make sure that a DB upgrade
is a noop if someone tries to redeploy the db on top
of an existing deployment.

Fixes bug 1298916

Change-Id: I17d1ccb3123346e9138ad1701fa5c6e1dd81529a