Allow as_dict with no assembly_id FK set in Component
Component.assembly_uuid used to raise an exception when no assembly_id
was set. It was an acceptable behavior but when you were calling as_dict
you were getting a ResourneNotFound and couldn't retrieve your dict.
The purpose of this patch is to allow the use of as_dict with no assembly
set as FK.

Change-Id: I3518e0dfb3f85631260474ba465146314e48a12c
Closes-Bug: #1306731