Fix error: "Stack already exists" in trigger workflow
On git push, the trigger workflow gets called.
It then then calls the deployer service which currently
tries to redeploy the assembly by creating a new heat
stack instead of updating the existing one. This
results in "Stack already exists" error.

The deployer code should first check if the stack for
the assembly already exists. If yes, it should update
the stack, else it should create a new stack.

Fixes bug 1311275

Change-Id: I2d4810361c1b63db0a17168331114cf17c832bd9