Enable Generic get_url_for method
Last version of keystone v3 API has a new format for the service
catalog. This resulted in Solum not working with a token generated
with the v3 API.

In this patch, we use a generic method to get the endpoint url
of a service, by calling keystone 'url_for' method. We also use
keystone client v3 by default now.

Fixes bug 1311542

Change-Id: Ib0866ab05ef8b072f486e6dc27be8fae8b640482