Fix devstack gate by increasing Swift max_header_size
Devstack gate is failing due to our new service catalog header
which is too large for swift to handle.

This patch increases this option in solum devstack setup as a
temporary solution. I will work to have this merged in devstack
itself and once it's done, we could revert this patch

Closes bug 1316585

Change-Id: I84c0ce321c0d4fc999e707f63be429b519b3a703