Docker Build-Script Enhancements
- Speeds up build times for ghost by 2X.
- Allows caching of git repos.
- Allows reusing images if repo is unchanged
- Git dir name is now a hash of the git URL.
- Added an elapsed time indication in build log.
- Simplified the image id fetch to use awk.
- Eliminated the redundant echo at the end.
- Corrected bug in previous patchset that gets an
incorrect image id from glance.

Build times taking 5 minutes before are now
completing in 160 sec. Allows caching the git
repo and will git pull if there is a previous one
present. If the git pull is up-to-date, then a
new image is not built, and the previous one
is used. This can be changed by setting the
REUSE_IMAGES_IF_REPO_UNCHANGED environment variable
to 0 to cause a rebuild even when the repo is
unchanged since the last build.

Enhances the Docker build script so that it keeps
timestamped log files. The log file location can
be controlled by setting the SOLUM_BUILD_LOG
environment variable.

Change-Id: Ie5523d81a49bb4d94291f35091435819b3bc5a05