Store heat_stack_id and component_type in component table
Added two attributes to component - heat_stack_id and
component_type.

When we need to update a heat stack (while updating
an assembly), the stack_id can be obtained from the
component table rather than querying the Heat API.
However, during assembly delete we still need to
query the Heat API to ensure that the stack is
actually deleted before deleting the assembly.

Closes-Bug: #1312856

Change-Id: I6ebf9b32d80d6c9f5f51e7f77d7c5ff344ef29ef