Use auth_token_info to initialize solum_keystoneclient sessions
Now we have the token info available in the request context, we can
avoid always re-requesting another token every time a client is
created.

This should be better for performance as it avoids one extra request to
keystone, and it also enables trust-scoped tokens to be more easily
used, since these cannot be used to request another token, thus
authentication will fail with the current scheme.

This is needed for the builder to be used from a mistral task.

Change-Id: Ic03e315a7a7fb29bd8a499032b61befea260ab7e
Partial-Bug: #1317293