Create an empty stack when creating the pipeline
When Heat creates a stack it needs to create a trust,
at the moment this can't be done in the context of a mistral task as
a trust token is passed into the task and creating a trust
from a trust token is not currently allowed.

This patch is a workaround to create the Heat stack with
a "normal" user token. Then the mistral task will only
do a stack update, which can be done with a trust token.

https://blueprints.launchpad.net/solum/+spec/pipeline
Change-Id: I23ce79761e3ebcd46cfbef810330ff6b7c9bde4e