Configurable base location of worker scripts
Depending on how Solum is installed, the contrib directory
may not exist in the directory three levels above handlers/shell.
In such a case, shell Handler's build and test functions  will fail
as it cannot find those scripts.

This change takes advantage of the Handler's proj_dir property,
allowing it to supercede the relative path if it's set.
The name proj_dir reflects the directory that contains the contrib directory,
among others, because that is where the original project codebase lies.

Change-Id: I9add45d93d5be801193e7ac45235cb55c347020b