Send back testing results
As a CI tool Solum should be able to report testing
results back to user. This patch handles the case of
hosting code on Github. By leveraging Github API,
Solum can send back the statuses of testing, i.e.
pending/success/failure and a log url for user to
inspect results.

A Github OAuth token is used for posting statuses.
The token will need to be generated by a user and
put into Solum plan file.

Closes-bug: 1365651

Change-Id: I223418f63d663a9673963b1f6a7a4367bd1766e1