Reverse diff check in check_uptodate
Check_uptodate creates what it thinks our example conf should look like,
and compares that to what we actually have. In case of an error, it shows
a diff of the two files, but the diff is effectively how "expected" was
changed to produce "actual".

By reversing the comparison order, the diff is then the changes to apply to
"actual" to get "expected". We can apply that diff with patch to quickly
resolve the problems check_uptodate found.

Change-Id: I805c2097ef271b403add81702638b5c75df4c25e