Check that public repos exist before cloning
Our unittest and build scripts use the git CLI to clone
and checkout repos. If a clone of a URL that appears good
results in a 404, git CLI will prompt at the terminal for
a username and password, assuming the repo was not found
because it is private. This causes solum-worker to freeze
indefinitely as it waits for this input.

This patch adds a function to common/utils to check via curl
that a repo is visible to the public, to keep the git CLI
from needlessly asking for credentials. Of course, this patch
also incorporates this utility into unittest-app and build-app.

Closes-Bug: 1375471
Change-Id: Ia133966a77b7f38b0a2617a482ab9223ab8f9b08