Change the hard coded build cmd in _run_unittest()
Use _get_build_command() to find the correct build script,
as a result the worker 'unittest' and 'build' APIs would have
the same signature, so removed unittest API and added
'verb' param to indicate which method to cast to. Also changed
the 'build' API to be a more general name, i.e. perform_action.

Change-Id: I532dd7df76f8bf1a857cad78e7561449e8d6a7be