Support private git repo's without using Barbican.
This patch adds support to run an instance of Solum
without using Barbican for storing private git repo secrets.
Secrets get stored on the host machines filesystem if the
config value is set to disable using Barbican.

This feature is great for two uses
1) Running internal instances of Solum when security is
not a big concern and the Ops\Dev team does not want
to run an extra OpenStack service.

2) Testing Solum in a vagrant environment without running
Barbican.

Change-Id: I6c674f9471ddfa641490f59b32fb2ea18facef82