Rename deployer delete_heat_stack to destroy
It doesn't make any sense that all deployers regardless of
their use of heat should have a method naming heat.
This patch renames the method to the more generic 'destroy',
and also adds the missing method to deployer's noop handler.
Without the method noop was in violation of its API.

Change-Id: Ia525678055a40474413dbe2f4a5a062da7d92121
Closes-Bug: 1350579