Support for building Custom languagepack
I'm extending the builder's Image API to allow
end users to build custom languagepack by passing
in a git repo url. The repo should contain a dockerfile

Here is a sample repo:
https://github.com/murali44/Solum-language-pack-Chef

This patch contains a new script 'build-lp', which
builds a docker image and uploads in to glance.
The glance image is tagged with 'Solum:lp' and
some metadata passed in(compiler version, runtime env, etc).

Change-Id: I75238cd580ac8c4d1fee4a60fa3fd61d86cafaa6