Remove solum.conf.sample from source control
We have a tool--literally tools/config/generate_sample.sh--
capable of creating a perfect sample config file at will.
Using it only to check that the sample in source control is right
doesn't make a lot of sense; I think if we trust its opinion so
strongly we ought to make it work for us.

If we decide we do want to keep a sample conf in source control,
I think we ought to manually generate it and keep it in doc/ with
a note about when it was last generated.

I've modified tox.ini to remove the check_uptodate job, as it's no
longer necessary. A lot of these config options are the result of
upstream dependencies whose updates are out of our control.

I've also modified our devstack lib script to generate the file
instead of copying it from the old place.

Change-Id: Ia204de5f4b571f68a3e5242eba2c0eafda7f0a5c