Introduce run_cmd in the planfile
I'm adding a run commnad to the planfile. This is the application
entry point and should be provided by the application developer.

I've also refactored the way test and run commands are passed to
the unit_test and build scripts. It was getting messy as we add
new variables.

This patch takes the run command from the plan file and passes it
to the build-app script. The command is not being used to start the
application code yet. That will be submitted in a different patch.

Change-Id: I9ef36b8f25a3b4c17fcd9fdec3d0e9e2423bf86f