Delete local languagepack image after building the DU
When building a DU based on a custom LP, we download the LP image
to the worker node to build the DU image on top of it.

Once the DU is built, we need to delete the LP from the worker,
else lingering LP images will cause the worker to run out of storage.

- Added a config option to activate this feature

Closes-bug: 1419067

Change-Id: If3fcebe494ff07c8076e0a23a098b10f8384e449