Create temp_url for LPs uploaded to swift
Languagepacks created today and stored in swift for one tenant are not
accessible by another tenant. This blocks the usecase where a developer
wants to use an operator created languagepack.

We need to change the way we store the reference to a
languagepack in the DB. If we create temp_urls for the LP in swift, it
can be downloaded by anyone who has access to the LP id.

Closes-bug: 1426577

Change-Id: I2ed1a3a5825df6530dc89ae0a35abc3e3b20c377