Check for code attribute in wrapped exceptions
When an unexpected exception is thrown in Solum,
the wrapping of those exceptions itself throws an error
because not every exception has the attribute 'code',
and the original exception is swallowed as the AttributeError
is surfaced.

By looking before assuming the exception has a code attribute,
we can properly surface the exception at least as a 500 for
further investigation.

Closes-Bug: 1453291
Change-Id: I754d956e6c72f4ae94c3ec3573674aee4f7df0bc