Admin users are able to access apps belonging to other tenants.
Users with an admin role (within a tenant) are able to list and
manipulate apps belonging to other tenants.

Users should be allowed access to apps that belong to the same
tenant as the user.

We need to introduce a global admin for use by customer service,
devops and devs for troubleshooting end user apps. This follows
Heats model of using a single tenant created specifically for
accessing apps belonging to all tenants.

Change-Id: I3524f47d051dd60fe3440b17f1574811f4cd1c65
Closes-bug: 1454838