Swift client wrapper
We currently upload logs and images to swift in two ways.
- Using the python swift client directly to upload logs
- Using a stand alone python script which is called from our bash
  scripts to upload images.

We need to consolidate these two so we have a common place for our
swift integration code which supports the following

- Upload and download objects.
- Check object size to make sure we are under the 5GB swift
  constraint.
- Retry logic
- Delete objects

I'm also changing the swift client code to always create a
non cached client. This is required for our upload\download
retry mechanism to work properly.

Change-Id: I9993d0b74197a517920468b7fd07e29bbc305ed0
Closes-bug: 1456807