Delete LP logs from swift
The LP delete method does not delete logs.

This results in orphaned log files being left in the users swift
account. The user gets charged for storage and it can be hard to
identify these log files for manual deletion by the user.

This patch deletes the log files from swift or local filesystem
(depending upon the config setting).

Change-Id: I8d799cc9a63851685766a7dec3bc4b806d688911
Closes-bug: 1457682