Fixing concurrency bug in heat deployer
Problem:
We want to ensure that only one assembly is in a READY state
for an app. This is to ensure that Solum does not keep around
more heat stacks than are necessary. If there are more than one
assemblies then there are that many number of heat stacks,
which we want to avoid.

So the solution that we came up with is to delete other
READY assemblies when a new assembly has become READY.

However, this solution as is currently implemented
has concurrency issues in the situation when there are multiple
deployers in the system. In our testing we have seen both of
the following:
(a) when two assembly create are triggered in quick succession both
    are deleted
(b) when two assembly creates are triggered in quick succession
    none is deleted

This patch fixes the concurrency issue to ensure that both the
above situations don't arise.

The fix for (a) is to use the 'created_at' timestamp when deleting
assemblies. The idea is to delete all the READY assemblies which
have created_at timestamp which is smaller than the newly created
assemblies' timestamp. The reason this is correct is because there
will be only one READY assembly which will have the highest created_at
timestamp (because we are using a single database). So this assembly
will end up deleting other assemblies which are READY and have a
smaller (more recent) created_at timestamp.

The fix for (b) is to perform the database writes to update an
assembly's status in a synchronous manner in the deployer itself.
Current code performs this update in an asynchronous manner via the
conductor. In a multiple deployer scenario this asynchronicity causes
race conditions which leads to situations like (b).

Closes-Bug: #1457200
Change-Id: I3482399fb08ebe19870f6dae69e2edc2a0c07685