Delete APP logs from swift
The APP delete method does not delete logs.

This results in orphaned log files being left in the users swift
account. The user gets charged for storage and it can be hard to
identify these log files for manual deletion by the user.

This patch deletes the log files from swift or local filesystem
(depending upon the config setting).

Change-Id: Ia733d6e96286ca7048376c99179aae66d420c4b8
Closes-bug: 1457682