Delete LP image files from swift
The LP delete method does not delete the docker image
file from swift.

This results in orphaned image files being left in the users
swift account. The user gets charged for storage and it can be
hard to identify these image files for manual deletion by the
user.

This patch deletes the these image files from swift, when the
user deletes an LP using the CLI.

Change-Id: I90fd7f51b26fd580ac3ee68c243a33ebb6671aaf
Closes-bug: 1459766