Modifying the heat deployer to use injected file to download DU
Problem:
We have noticed that occassionally the downloading of DU on the
VM will fail to download the entire file. This causes failure
during docker load, which loads and makes the DU available to
be run on the VM.

Solution:
One way to address this issue is to check docker load for success,
and try the steps of downloading the DU and loading it again, if it
has failed. Towards this we have developed a script which will
perform such checks. We inject this script into the VM using the
mechanism of 'get_file' provided by Heat.

This patch implements this solution.

As a consequence of this change, the heat deployer is simplied,
as we no longer need to customize the template in our code.
We just need to pass the correct parameters to the template.

Change-Id: I14c4729727b17ab7671c155a1ed32765af1e7b6b