Refactoring heat deployer's multi-port checking code
The code which checks whether all the ports of the DU are active
was coupled with the code that ran the check for du_attempts
number of time. This coupling was incorrect. If the du_attempts
was less than the number of ports, we would have ended up not
checking all the ports.

This patch fixes this situation. We decouple the checking of the ports
from trying the check du_attempts times. We also maintain a list of
successful_ports so that in each du_attempt we only check those ports
which have not yet been successfully reachable.

Change-Id: Ib82e5547d9ab2184df3b3e3800301925d5682284