Support operator LP download with swiftclient
Operator LPs are only accessible using a public url today.
We currently 'wget' these LP images while building a users
application image on top of the operator LPs.

This works fine with small images, but as docker images get larger,
I've noticed the images get corrupted due to network issues and
not downloading the images in chunks.

We need to also support downloading the LP images from a swift
container belonging to the service operator using a swift client.
Swift has already implemented retry logic, downloading in chunks
and it faster as well.

Change-Id: I1ad43c5122516e9bbb9507534c842ab0c0e571ef
Closes-bug: 1466233