Fix for bug which leaves logs in swift after App deletion
During app deletion, we look for an image id which points to the
DU which needs to be deleted. If the app failed in the unit test
phase, we don't have a DU yet and the deletion code returns due
to an exception, thus failing to continue with deletion of logs.

We should be deleting the unit test phase logs.

Change-Id: I78befbcfd8605b25371597e4fc07f0908b11730d
Closes-bug: 1463986