Setting default values for languagepack admin creds
Currently the default values are not set (empty). Due to this
languagepack build fails in vagrant/devstack setup.
So setting the default values that are appropriate from vagrant/devstack
setup point-of-view.

In real deployment these creds should be set to the lp admin user
that is created in the system

Change-Id: I55700b00b06e99184f8bfad1128d6208f5ff7ad7