Correctly parsing the server ip from the created heat stack.
The heat stack-show command generates a list of two entries for a stack's
output. One entry is for the output_key "URL" and other is corresponding to
the output_key "public_ip". We want to parse the IP address of the server
by parsing the entry corresponding to the "public_ip".
Currently we are picking the first entry from the outputs.

This is incorrect, as the order of entries in the output may not be fixed.

Change-Id: Iaf09f670e87097c85f3363e03dddac84cf4db497
Fixes-Bug: #1485750