Parsing subnet-id for ipv4 subnet from neutron
Parsing the subnet-id from neutron which corresponds to the ipv4 address.
Currently, when parsing the neutron subnet ids, we are picking the first
entry from a list. See line 26 of:

https://github.com/stackforge/solum/blob/master/solum/common/heat_utils.py

However, the first entry could be either for ipv4 subnet or ipv6 subnet.
We want to pick the ipv4 subnet id.

Fixes-Bug: #1485749

Change-Id: I6ff79853e9879436dbbb730eda637179be2082e0