Adding workflow objects and models
While we are introducing the workflow resource and related objects, we
want to ensure that we remain backwards compatible with the assembly
and plan models which are currently used quite extensively internally
within Solum. Especially, the worker/shell.py and deployer/heat.py make
heavy use of the assembly abstraction.

With this patch we are adding a weak one-to-one association between
workflows and assemblies.

We have also introduced the workflow_table and have added some required
fields to the app table.

Change-Id: I3eed8f3515990b460bfeb3db07dac484edf543c8
Partial-bug: #1486200
Co-Authored-By: ed--cranford (ed.cranford@gmail.com)