Added workflow handler and connected it to worker and deployer
Currently Solum uses notion of assembly quite extensively in
worker/shell.py and deployer/heat.py. While we are introducing
the new workflow resource and related objects, we want to ensure that
we can start using the workflow resource without having to wait for
complete overhaul of worker and deployer.

So to achieve this goal we have introduced an apapter class
in the workflow_handler. The adapter class provides a method to
create a plan and an assembly. This method is invoked upon each
POST to workflow. A one-to-one association is maintained between
each workflow and the corresponding created assembly.

Finally, to complete the end-to-end deployment we have introduced
a method in the deployer/heat.py to update the 'app' corresponding
to the invoked workflow with the status and the url of the
deployed application.

Change-Id: Ia08bf55f91d16ecc47e37b9dc1e8bae96a8141c3
Partial-Bug: #1486200
Co-Authored-By: ed--cranford (ed.cranford@gmail.com)