Fixes to work with WSME 0.8.0
WSME 0.8.0 adds strict data validation checking. This includes
disallowing any attribute which is not defined in the API data
model to be passed in POST/PUT calls.

We were passing extra attributes, such as trigger_uri for
assembly creation, base_url for language_pack creation,
and plan_uri for component creation. Due to this the tests,
and cli-based user interactions are failing.

This patch fixes above three issues.

For assembly creation tests, we remove the trigger_uri from
the test data as we don't need to pass it anyways.

For language_pack creation, we add the base_url to base types.
The base_url is passed as part of the request body for POST/PUT
when we use the solum cli. Internally this is happening due to
solum cli using the OpenStack common cli which treats the
base_url as one of the keyword arguments and sends all the
kwargs as part of the POST/PUT body.

So probably the right fix here is to modify the OpenStack
common cli to not pass base_url as part of the body. But until
that is done, a workaround is to add the base_url to the list
of attributes to be expected on the API side. That is what we
are doing.

For component creation tests, we add the plan_uri to the component
data model. We are not familiar with the CAMP component data
model to make a call of whether adding a plan_uri to the data
model makes sense or not. I tried removing the plan_uri from
the data that is sent as part of the tests.
That fixed few failing tests, but there
were at least three which were still failing. I have not looked
in details why those remaining tests are failing. But adding
plan_uri on the component API data model seems to make all
tests to pass. So that is what we are doing.

Closes-Bug: #1491504

Change-Id: I7bc58a2fca41ace67b9c1fad6918c2012eb4f2fe