Changing glance commands to use version flag
Version 2 of Glance API has been released and we need to decide
whether to migrate to using this new version or continue using
version 1.

Version 2 requires passing in more flags
than we are currently using in our glance commands. In order
to pass these additional flags we have to modify worker/shell.py.

Using version 1 requires passing in a single version flag with
the glance commands.

This patch chooses the option of using version flag to keep using
version 1 as we know that Solum works with version 1.
To migrate to version 2, we will have to do additional testing
We will do that in a separate patch.

Change-Id: If38a56957c591998c4459b700624e12cc6dd4417
Partial-Bug: #1499520