Upgrading tempest-lib to unblock dvsm gate
tempest-lib has changed signature on one of the methods.
The changed method signature is available in tempest-lib-0.11.0.
We cannot wait till that version makes it to global-req.
So installing that version in post_test_hook.sh.

Also, it seems that we need to setup the tempest.conf file
as required otherwise tempest uses default values, which are not
appropriate for our testing. So setting up tempest.conf as well
in post_test_hook.sh

Closes-Bug: #1509407
Change-Id: I4733604af4effe904bbadca5e9083d03fc082d93