Removing incorrect barbican setup in pre_test_hook
Removing special setup that we were doing for Barbican on the
devstack gate. Two reasons for this. First, the setup code has
drifted now that lib and extras.d dirs have been deprecated in
Devstack. Second, we are not using Barbian right now so there is
really no use of this setup. If and when we do decide to integrate
Barbican again, we can consider configuring the devstack gate with
any required Barbican setup.

Closes-Bug: #1512859

Change-Id: Id39dd9a08d4a813f0fb28a2c108d3820b46f49f9