Add scale_config column to the app table
Adding a column named scale_config to the app table.
The idea is to use it to store an app's scaling related
configuration. The type of the column is set to text, so that
we can store any required information in this column.
For now, the plan is to store following info in json format:

{<app-name>: {target: m, current: n, du: du-name/du-id}}, where

app-name is the app-name

m is the number of target instances requested in the most
recent scaling operation.

n is the number of actual instances running after the most
recent scaling operation has been completed.

n <= m relation will always hold.

du-name/du-id is the name/id of the application du that
should be used in the scaling operation.

This column will also allow us to support multi-service
applications in the future. Lets say we have an app with
two services - web and backend. The scaling configuration
for both can be stored as follows.

{web: {target: w_m, current: w_n, du: w_du},
 backend: {target: b_m, current: b_n, du: b_du}
}

Change-Id: I94e41c0953faa1f5d51defa3d4ccf33384f58135
Closes-Bug: #1541467