Added devstack-provenance file
Added devstack-provenance file that contains
hashes and versions of all the services and
packages in devstack that led to successful
app deployment.

If your devstack environment is not deploying
applications correctly then you can use this
file to find the versions of services and packages
that have worked before. You can then pin different
services and packages to versions identified here
to get a working devstack environment.

Update this file only after a successful
app deployment in devstack

Change-Id: I86c19bd5b1c9c7b482cd1eaa8b4ea31b27f40303