Moved CORS middleware configuration into oslo-config-generator
The default values needed for solum's implementation of cors
middleware have been moved from paste.ini into the configuration
hooks provided by oslo.config. Furthermore, these values have been
added to the default configuration parsing. This ensures
that if a value remains unset in solum.conf, it will be set
to use sane defaults, and that an operator modifying the
configuration file will be presented with a default set of
necessary sane headers.

Change-Id: I6f30224ac1b11fc4019dbc5ae5ec1e1fedbfe97d
Closes-Bug: 1551836