Devstack plugin update
- Installing docker and nova-docker driver in plugin.sh
- Handling virt driver option in /etc/nova/nova.conf
  differently for local dev env and the gate

For local dev environment (vagrant, vagrant+devstack),
we want the the virt driver option to be set to 'docker'.
So doing that in plugin.sh. On the other hand, our devstack
gate is not configured to use nova-docker. It is configured
to use 'vm'. Therefore, unsetting the virt driver option
in the post_test_hook.sh

Change-Id: I5349ecee1dec3a56a8836ee97a238b3776748e6b