Copy nova docker from the new location
The 'nova-docker' driver is an experimental driver in Zun and
it has been moved to /contrib folder. Let's copy nova-docker driver
from the /contrib folder to work around.

In long-term, Solum should switch to the 'docker' driver which is the
default driver in Zun. This driver is well maintained by Zun team and
should be a better choice for Solum.

Change-Id: I27cccf6eb2dd90c69ab316adcc9eb409316ecf66