Fix lower-constraint deps handling
When the lower-constraints tox target was added, it was assumed the
install_command was just running the install and that the dependencies
and constraints were being set using "deps = ".

This fixed the install_command and deps to follow the expected pattern
so the lower-constraints job actual does install the lower constraints.

Also raises some lower-constraints that were found to fail once the job
was working.

Change-Id: I181f724626f5f5c49af2a8d1a5ed6632cc274858
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>